+++ packages/superposition/src/index.ts
@@ -1,5 +1,5 @@
-export * from './Mock/MockChrono';
-export * from './Mock/MockSuperposition';
+export * from './mock/MockChrono';
+export * from './mock/MockSuperposition';
 export * from './Plan/index';
 export * from './Chrono';
 export * from './ISuperposition';
