+++ packages/unscharferelation/src/Plan/__tests__/RecoveryEpoquePlan.spec.ts
@@ -1,4 +1,4 @@
-import { MockEpoque } from '../../Mock/MockEpoque';
+import { MockEpoque } from '../../mock/MockEpoque';
 import { RecoveryEpoquePlan } from '../RecoveryEpoquePlan';
 
 describe('RecoveryEpoquePlan', () => {
