+++ packages/unscharferelation/src/index.ts
@@ -1,6 +1,6 @@
+export * from './mock/MockEpoque';
+export * from './mock/MockUnscharferelation';
 export * from './Epoque';
-export * from './Mock/MockEpoque';
-export * from './Mock/MockUnscharferelation';
 export * from './Plan/index';
 export * from './Unscharferelation';
 export * from './UnscharferelationError';
